Revert "#000: Chauhan: Add custom decoder for eqcoord"
