+++ csw-params/shared/src/main/scala/csw/params/core/formats/ParamCodecs.scala
@@ -1,21 +1,20 @@
 package csw.params.core.formats
 
 import java.lang.{Byte => JByte}
-import csw.params.commands.*
-import csw.params.core.formats.EqCoordCodecHelpers.WireModel
+
+import csw.params.commands._
 import csw.params.core.generics.{KeyType, Parameter}
-import csw.params.core.models.Coords.*
-import csw.params.core.models.*
+import csw.params.core.models.Coords._
+import csw.params.core.models._
 import csw.params.core.states.{CurrentState, StateName, StateVariable}
 import csw.params.events.{Event, EventName}
 import csw.prefix.codecs.CommonCodecs
 import csw.time.core.models.{TAITime, UTCTime}
-import io.bullet.borer.*
+import io.bullet.borer._
 import io.bullet.borer.derivation.CompactMapBasedCodecs.deriveCodec
 import io.bullet.borer.derivation.MapBasedCodecs
 import io.bullet.borer.derivation.MapBasedCodecs.deriveAllCodecs
 
-import scala.annotation.nowarn
 import scala.collection.mutable.{ArraySeq => ArrayS}
 import scala.reflect.ClassTag
 
@@ -57,15 +56,7 @@ trait ParamCodecsBase extends CommonCodecs {
   implicit lazy val angleCodec: Codec[Angle]               = deriveCodec
   implicit lazy val properMotionCodec: Codec[ProperMotion] = deriveCodec
 
-  lazy val coordCodecValue: Codec[Coord] = {
-    implicit lazy val wireModelCodec: Codec[WireModel] = MapBasedCodecs.deriveCodec
-    @nowarn
-    implicit lazy val eqCordCodec: Codec[EqCoord] = Codec.bimap[WireModel, EqCoord](
-      EqCoordCodecHelpers.fromEqCord,
-      EqCoordCodecHelpers.toEqCord
-    )
-    deriveAllCodecs
-  }
+  lazy val coordCodecValue: Codec[Coord] = deriveAllCodecs
 
   implicit lazy val utcTimeCodec: Codec[UTCTime] = deriveCodec
   implicit lazy val taiTimeCodec: Codec[TAITime] = deriveCodec
