+++ csw-params/shared/src/main/scala/csw/params/core/models/Angle.scala
@@ -455,3 +455,40 @@ object Angle {
   /** round circle which marks degrees */
   val DEGREE_SIGN: Char = '\u00B0'
 }
+
+/**
+ * Java API for creating an Angle instance.
+ */
+object JAngle {
+  import Angle.*
+
+  /**
+   * Creates an Angle instance from the given value in degrees
+   */
+  def degree(d: Double): Angle = d.degree
+
+  /**
+   * Creates an Angle instance from the given value in arcMinutes
+   */
+  def arcMinute(d: Double): Angle = d.arcMinute
+
+  /**
+   * Creates an Angle instance from the given value in arcSecs
+   */
+  def arcSec(d: Double): Angle = d.arcSec
+
+  /**
+   * Creates an Angle instance from the given value in arcHours
+   */
+  def arcHour(d: Double): Angle = d.arcHour
+
+  /**
+   * Creates an Angle instance from the given value in radians
+   */
+  def radian(d: Double): Angle = d.radian
+
+  /**
+   * Creates an Angle instance from the given value in mas (milliarcseconds)
+   */
+  def mas(d: Double): Angle = d.mas
+}
